Add files via upload
Instructions for installing HamClock on a Raspberry Pi Zero 2 W and running it 'headless' so that it can be accessed by any other computer on your wifi network using a Web Browser. This provides an inexpensive way of using HamClock and the flexibility of operating and interacting with HamClock from any web browser.